requirements.txt
pip==21.2.3
PyPDF2==1.26.0
google-api-python-client==2.19.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
selenium==4.13.0